Refactor(family): dto의 데이터량 관련 필드 타입을 Long으로 통일

+++ src/main/java/com/pooli/family/domain/dto/request/UpdateSharedDataThresholdReqDto.java
@@ -9,5 +9,5 @@ import lombok.Setter;
 @Schema(description = "가족 공유 데이터 사용량 임계치 수정 요청 DTO")
 public class UpdateSharedDataThresholdReqDto {
     @Schema(description = "새로운 임계치 값(단위: Byte)", example = "1500")
-    private Integer newFamilyThreshold;
+    private Long newFamilyThreshold;
 }
